--- %ProgramData%/Microsoft/Windows/UUS/State/_active.uusver
@@ -1,1 +1,0 @@
-1023.910.1172.0
